Declare fields final
As the `ScanData` class is essentially a data struct all fields are
declared `final`. This make it clear the intent as a data struct. By
While the `scanRecord` is still mutable declaring the fields as `final`
starts the process of making this data struct a simple immutable
structure.

The packet detector just needs a simple set to track the data. By
declaring the field `final` and including the nullability annotation we
can guarantee that the field will never be `null`.